+++ v8go.cc
@@ -808,7 +808,7 @@ ValuePtr NewValueIntegerFromUnsigned(IsolatePtr iso, uint32_t v) {
   return tracked_value(ctx, val);
 }
 
-RtnValue NewValueStringFromByteArray(IsolatePtr iso, const uint8_t* v, int len){
+RtnValue NewStringFromBytes(IsolatePtr iso, const uint8_t* v, int len){
   ISOLATE_SCOPE_INTERNAL_CONTEXT(iso);
   TryCatch try_catch(iso);
   RtnValue rtn = {};
@@ -979,6 +979,7 @@ RtnString ValueToString(ValuePtr ptr) {
   // TODO: Consider propagating the JS error. A fallback value could be returned
   // in Value.String()
   String::Utf8Value src(iso, value);
+  printf("___ valuetostring %d \n", src.length());
   char* data = static_cast<char*>(malloc(src.length()));
   memcpy(data, *src, src.length());
   rtn.data = data;